backend/requirements.txt
fastapi
uvicorn[standard]
pandas
numpy<2.0
scikit-surprise
lightfm